// src/constants/enums.ts
export enum userVerificationStatus {
  Unverified = 'Unverified',
  Verified = 'Verified',
  Blocked = 'Blocked'
}

export enum tokenType {
  AccessToken = 'AccessToken',
  RefreshToken = 'RefreshToken',
  ForgotPasswordToken = 'ForgotPasswordToken',
  EmailVerificationToken = 'EmailVerificationToken'
}